Fix an incorrect setting name (SMTP authentication) (#2175)
* Fix an incorrect setting name

In the smtp yaml file, the setting is called 'authentication', however in config/environments/ it is looked up as 'plain'.
Fix this issue by correcting the look-up to use the correct name.

This solves an issue with e-mail when the SMTP server requires a different authentication method, such as cram_md5.

(Note it appears the name plain may come from a time at which the setting was a boolean plain)

* Update the documentation for the mail server authentication option in production.rb

(cherry picked from commit 13c0816192be81f714b29bdce36e02f7ee3e43fc)